refactor : PR 리뷰 반영
- ApiError 클래스 추가, statusCode로 에러 분기
- parseInt || 폴백 → isNaN 체크로 수정 (0 입력 버그 수정)
- uploadImage try-catch 추가
- birth-date 유틸 함수 utils/ 폴더로 분리
- 온보딩 상수 constants/ 폴더로 분리 (hiking, exercise)
- useSubmitOnboardingFromStore store/ 폴더로 이동
- OptionButton 반환 타입 React.ReactNode로 변경
- setProfile hook 컴포넌트 최상단으로 이동
