+++ features/onboarding/components/onboarding-exercise-detail-screen.tsx
@@ -1,48 +1,19 @@
 import { OptionButton } from "@/components/option-button";
-import { useSubmitOnboardingFromStore } from "@/features/onboarding/hooks/use-submit-onboarding";
+import {
+  DURATION_OPTIONS,
+  EXERCISE_LABELS,
+  FREQUENCY_OPTIONS,
+  ExerciseDuration,
+  ExerciseFrequency,
+} from "@/features/onboarding/constants/exercise";
 import { useOnboardingStore } from "@/features/onboarding/store/onboarding-store";
+import { useSubmitOnboardingFromStore } from "@/features/onboarding/store/use-submit-onboarding-from-store";
+import { ApiError } from "@/lib/api";
 import { toast } from "@/store/toast.store";
-import { RegisterOnboardingRequest } from "@/types/api.generated";
 import { router } from "expo-router";
 import { useState } from "react";
 import { Pressable, ScrollView, Text, View } from "react-native";
 import { useSafeAreaInsets } from "react-native-safe-area-context";
-
-type ExerciseDuration = NonNullable<
-  RegisterOnboardingRequest["exerciseDuration"]
->;
-type ExerciseFrequency = NonNullable<
-  RegisterOnboardingRequest["exerciseFrequency"]
->;
-type ExerciseType = RegisterOnboardingRequest["exerciseType"];
-
-const EXERCISE_LABELS: Record<ExerciseType, string> = {
-  GYM: "헬스",
-  HIKING: "등산",
-  RUNNING: "러닝",
-  SWIMMING: "수영",
-  HOME_TRAINING: "홈트레이닝",
-  PILATES_YOGA: "필라테스/요가",
-  SPORTS: "스포츠",
-  CROSSFIT: "크로스핏",
-  WALKING: "워킹",
-  NONE: "운동 안 함",
-};
-
-const DURATION_OPTIONS: { label: string; value: ExerciseDuration }[] = [
-  { label: "4시간 이상", value: "OVER_4H" },
-  { label: "2~4시간", value: "HOUR_2_4" },
-  { label: "1~2시간", value: "HOUR_1_2" },
-  { label: "1시간 미만", value: "UNDER_1H" },
-];
-
-const FREQUENCY_OPTIONS: { label: string; value: ExerciseFrequency }[] = [
-  { label: "거의 매일", value: "DAILY" },
-  { label: "주 3~4회", value: "WEEK_3_4" },
-  { label: "주 1~2회", value: "WEEK_1_2" },
-  { label: "월 1~2회", value: "MONTH_1_2" },
-  { label: "연 1~2회", value: "LESS_THAN_MONTH_1" },
-];
 
 export function OnboardingExerciseDetailScreen(): React.JSX.Element {
   const insets = useSafeAreaInsets();
@@ -67,8 +38,7 @@ export function OnboardingExerciseDetailScreen(): React.JSX.Element {
       await submit();
       router.replace("/(tabs)");
     } catch (e) {
-      const message = e instanceof Error ? e.message : "";
-      if (message.includes("API Error: 409")) {
+      if (e instanceof ApiError && e.statusCode === 409) {
         toast.show("이미 등록된 사용자입니다.");
         router.replace("/(tabs)");
       } else {
