+++ features/onboarding/hooks/use-submit-onboarding.ts
@@ -18,15 +18,3 @@ export function useSubmitOnboarding() {
   });
 }
 
-export function useSubmitOnboardingFromStore() {
-  const toRequest = useOnboardingStore((s) => s.toRequest);
-  const { mutateAsync, ...rest } = useSubmitOnboarding();
-
-  async function submit(): Promise<void> {
-    const request = toRequest();
-    if (!request) return;
-    await mutateAsync(request);
-  }
-
-  return { submit, ...rest };
-}
