+++ utils/birth-date.ts
@@ -1,0 +1,14 @@
+export function formatBirthDate(text: string): string {
+  const digits = text.replace(/\D/g, "").slice(0, 8);
+  if (digits.length > 6) {
+    return `${digits.slice(0, 4)}-${digits.slice(4, 6)}-${digits.slice(6)}`;
+  }
+  if (digits.length > 4) {
+    return `${digits.slice(0, 4)}-${digits.slice(4)}`;
+  }
+  return digits;
+}
+
+export function isValidBirthDate(date: string): boolean {
+  return /^\d{4}-\d{2}-\d{2}$/.test(date.trim());
+}
